requirements.txt
amazoncaptcha==0.5.2
fake-useragent==0.1.11
openpyxl==3.0.9
pandas==1.3.4
playwright==1.17.2
playwright-stealth==1.0.5
PyYAML==6.0
Scrapy==2.4.1
scrapy-fake-useragent==1.4.4